+++ huntflow_api_client/entities/organization_settings.py
@@ -1,0 +1,35 @@
+from huntflow_api_client.entities.base import BaseEntity
+from huntflow_api_client.models.response.organization_settings import (
+    CloseReasonsListResponse,
+    HoldReasonsListResponse,
+)
+
+
+class OrganizationSettings(BaseEntity):
+    async def get_hold_reasons(self, account_id: int) -> HoldReasonsListResponse:
+        """
+        API method reference
+            https://api.huntflow.ai/v2/docs#get-/accounts/-account_id-/vacancy_hold_reasons
+
+        :param account_id: Organization ID
+        :return: List of vacancy hold reasons
+        """
+        response = await self._api.request(
+            "GET",
+            f"/accounts/{account_id}/vacancy_hold_reasons",
+        )
+        return HoldReasonsListResponse(**response.json())
+
+    async def get_close_reasons(self, account_id: int) -> CloseReasonsListResponse:
+        """
+        API method reference
+            https://api.huntflow.ai/v2/docs#get-/accounts/-account_id-/vacancy_close_reasons
+
+        :param account_id: Organization ID
+        :return: List of vacancy close reasons
+        """
+        response = await self._api.request(
+            "GET",
+            f"/accounts/{account_id}/vacancy_close_reasons",
+        )
+        return CloseReasonsListResponse(**response.json())
